add undefined is_monotonic attribute to Index
This attribute is used in some algorithms to be faster when the index is
known to be sorted.  In our case we don't care to signal this because
those algorithms are unlikely to be faster on the GPU.  We don't mind
doing the full-pass algorithms every time.

However, this may also confuse users, who might check this and assume
that their index is not sorted, even if it is.  I've gone with the value
`None` here sometimes used with ternary logic to signify that we're not
making a claim either way.

This is used by dask.dataframe when doing repartitioning.