Include `strides` in `__cuda_array_interface__`
As some implementations of `__cuda_array_interface__` expected `strides`
to be filled, go ahead and provide the `strides` field. As we ensure the
data is contiguous, we know the step size is just the size of the item
in the columns. Additionally we know columns are 1-D. So just make the
`strides` a singleton `tuple` with the item size.