Reorganize `data` type checks and exceptions
Turn the unexpected `data` type case into a flat out exception. Make a
case for `data` being an address where `size`'s type is also checked.
Should make this a bit clearer on what the cases are and when the
exceptions occur. Also should simplify the process for adding new type
cases.