Use "type-serialized" for pickled types like Dask
Dask uses the "type-serialized" header to store pickled type
information. Not to mention the "type" header is often used by Dask to
mean something different. As a result Dask could potentially stomp on
the information we are storing in the "type" header. So instead of
deviating from Dask, just store the pickled type information in
"type-serialized" as well. This should make it easier to more tightly
integrate our serialization with Dask's down the road.