Add Dask serializers for cuDF objects
In the event that serializing CUDA objects directly is not possible,
this performs the next best thing, which is to serialize objects using
Dask's serialization protocol. The protocol requires that data be on the
host in 1-D contiguous `memoryviews`. So we perform serialization as we
otherwise would. As a last step we perform a device-to-host transfer of
the frames. Then we hand this off to Dask to serialize. When
deserializing the data, all of the deserializers already work as frames
are turned into `Buffer`s, which perform a host-to-device transfer if
needed. This provides us an option that avoids pickling. As a result we
are able to serialize things with Dask more efficiently using this
protocol.