Serialize everything with CUDA first then run D2H
To simplify the code paths a bit, this just dispatches to
`cuda_serialize_cudf_object` at the beginning of
`dask_serialize_cudf_object` instead of calling `.serialize()`. This
ensures any checks we perform on that result are also performed in the
host serialization path. Also makes it a bit clearer that host
serialization is just device serialization followed by a device-to-host
transfer.