Support out-of-band pickling when available
In Pickle's protocol 5, out-of-band buffers are supported, which avoids
unnecessary copies when serializing data. In other words, this similar
to Dask's custom serialization except for pickling and can be supported
by any library that can use this Python standard. The only requirement
is we wrap any bytes-like objects in `PickleBuffer`s in our
`__reduce_ex__` method, which is what we do here. If an older Pickle
protocol is in use, we simply skip this path and go about pickling the
NumPy array as we would have otherwise.