Split Dask deserialization methods by dask/cuda
As Dask already routes to the right deserialization function based on
whether `"dask"` or `"cuda"` deserialization was used, split the
deserialization function into 2 deserialization functions for these 2
cases. This should avoid us needing to check which deserialization to
perform. Also this is very lightweight to implement thanks to the fact
that this logic is already captured in methods of `Serializable`.