Replace index type-dispatch call with indexalator in cudf::gather (#6457)
The `cudf::gather` API performs a double type-dispatch. The first one is against the index column parameter which will have an integer type like `int8, int16, int32, int64, uint8, uint16, uint32, or uint64`. The second dispatch is for the gather column data itself. This means the compiler will generate 8 copies of the gather code (second dispatch) where the only difference is the memory access to the index data. 

Using the index normalizing iterator (indexalator) introduced to help with dictionary support for the first dispatch greatly reduces the code size and compile time for `gather.cu`. The compile time was reduced on my system from 200s to 40s. The object file size changed from 10.8MB to 4.2MB. This also resulted in a size decrease for `libcudf.so` of 4MB.

Also, running the `gbenchmark/GATHER_BENCH` showed no significant change in performance using the indexalator over the type-dispatch. The following is the result of comparing `GATHER_BENCH` before and after replacing the first type-dispatch with the indexalator.
![gather-benchmark](https://user-images.githubusercontent.com/45795991/95337525-eb828380-087f-11eb-8e60-fd9fb4a83a3d.png)
The chart shows the difference in values in the outputs. The negative values indicate the indexalator output was faster/better while the positive ones were slower/worse.