Fix subword tokenizer metadata for token count equal to max_sequence_length (#6621)
Fixes #6618

The final metadata value calculation was incorrect when the number of tokens produced equaled the `max_sequence_length`. The logic mistakenly fell through to the truncation/stride calculation instead. This would work only for specific ranges for the `stride` value. Regardless, if the number of tokens fit, the stride value should not be pertinent.

This PR fixes this and cleans up the logic for setting the metadata vector values. Also added a gtest for this specific case.