[REVIEW] Fix csv writer handling embedded comma delimiter (#6643)
Closes #6641

The csv writer_impl logic is checking for the ',' delimiter using already freed temporary device memory.

The ctor for predicate_special_chars accepts a string_scalar object but is passed a std::string.
The implicit conversion creates a temporary string_scalar correctly.
The ctor then sets a string_view of the string_scalar data as a member variable. This is just a view and does not make a copy of the data inside of string_scalar.
This means that once the ctor returns, the temporary string_scalar is destroyed and the string_view is now pointing to freed memory.

The PR creates a string_scalar before the ctor and then passes a string_view instance to the predicate_special_chars ctor instead. This also removes the need to pass a stream parameter to the ctor. Now the string_scalar object has the same lifespan as the predicate_special_chars instance.