[REVIEW] Fix integer parsing in CSV and JSON for values outside of int64 range (#6446)
Fix #6314, #6449 and #6274.

An integer encountered in a column can belong to one of the three categories :

    Negative small integer - value ranges from -1 to int64_min
    Positive small integer - value ranges from 0 to int64_max
    Big integer - value ranges from int64_max+1 to uint64_max
    Out of range - values that are too big for uint64 or too small for int64

    If a column contains a mix of negative and positive small integers then the column is assigned to be of dtype int64.
    If the column contains a mix of positive small integers and big integers then the column is assigned to be of dtype uint64.
    If the column contains a mix of negative small integers and big integers then the column is assigned to be of string type.
    If a column has an integer that cannot be expressed as either uint64 or int64 then the column is assigned to be of string type.

Some checks are also added to correctly handle the parity sign and leading zeros. For example, if an integer is expressed as -0000, then it is still counted as a positive small integer. After removing leading zeros from an integer if we find that the number of digits is the same as int64_min, int64_max or uint64_max then it is lexicographically compared against the appropriate edge value.