Fix cuDF benchmarks build with static Arrow lib and fix rapids-compose cuDF JNI build (#6732)
The cmakelists for cuDF benchmarks (which are not built by default) failed to compile when
ARROW_STATIC_LIB was enabled. This fixes that failure.

Also adds a cmake variable passthrough to the cuDF JNI pom.xml to enable building
cuDF JNI bindings when the cuDF build directory is not the default of cudf/cpp/build,
as is common when using rapids-compose to build from source.