Fix hash join hash values mapping to reserved empty value (#6735)
Closes #6730

The hash join implementation stores the hash value of the row as the key in the hash table. There were cases where that hash value could end up equal to the reserved "empty" key value.

This PR remedies the situation by remapping any hash value that ends up equal to the empty sentinel to an arbitrary, new value (empty - 1 to be exact). This technically increases the number of hash collisions at that value, but this shouldn't be a problem in practice.