Fix race conditions in parquet (#6766)
closes #6244

This PR addresses one of the potential race condition with data_start which is being read and updated in different warps.

I have covered all except `unsnap.cu`, which is using `__nanosleep` with consumer/producer approach to handle synchronization. But this doesn't seems be recognized by tool and end up generating False Positives. And in some other places, `__shfl_sync` and `__ballot_sync` are being used which are not recognized as synchronization points by cuda-memcheck. So skipping these as False Positives.