Move template param to member var to improve compile of hash/groupby.cu (#6835)
This change reduces the compile time for hash/groupby.cu from 16 minutes to 9 minutes.
The `skip_rows_with_nulls` template parameter is set to avoid calling (and inlining) `cudf::bit_is_set()`. This function is minimal compared to the `cudf::detail::aggregate_row` function that must be inlined twice to accommodate this template parameter. Simply changing this to a member variable means we still do not incur an extra call to `cudf::bit_is_set()` when appropriate but also means we generate half as much device code for this specific function. The `cudf::detail::aggregate_row` code is quite significant.