Add null count test for apply_boolean_mask(#6903)
This PR adds a test to exercise the issue described in #6733. This issue was only reproduced on a laptop Pascal GPU, but I think it's a good test to have.  In summary, `copy_if`, used by `apply_boolean_mask` computes the output null count during as part of its custom scatter kernel, rather than using `cudf::count_unset_bits`.  #6733 describes an issue where the former is different from the latter. So it's good to have a test that verifies they get the same null count.

And since it's difficult to get a repro on a similar machine, this is a first step.

Authors:
  - Mark Harris <mharris@nvidia.com>
  - Keith Kraus <kkraus@nvidia.com>

Approvers:
  - Karthikeyan
  - Devavret Makkar

URL: https://github.com/rapidsai/cudf/pull/6903