Correct the sampling range when sampling with replacement(#6884)
This corrects an issue with the sampling range used when replacement=True. Before, it sampled the range 0 through `num_rows` meaning it could sample `num_rows` even though it's one position out of bounds. This caused sample to return values not present in the original DataFrame.

I also created exceptions for sampling on empty DataFrames that match pandas, as well as an exception for sampling when `axis=1` and `replace=True` as cudf does not support DataFrames with duplicate columns.

This closes #6532

Authors:
  - Chris Jarrett <cjarrett@dt08.aselab.nvidia.com>
  - Mark Harris <mharris@nvidia.com>
  - ChrisJar <chris.jarrett.0@gmail.com>

Approvers:
  - Keith Kraus
  - Mark Harris

URL: https://github.com/rapidsai/cudf/pull/6884