JNI support for creating struct column from existing columns and fixed bug in struct with no children(#7084)
The primary goal of this is to add in java APIs to create a struct column from other existing columns.  As a part of this work I found a very small bug in the column_vector constructor that copies data from  a column view for a struct column with no children in it.  Spark supports this use case so I thought it would be good to test/fix the issue.

Authors:
  - Robert (Bobby) Evans <bobby@apache.org>

Approvers:
  - Kuhu Shukla (@kuhushukla)
  - Vukasin Milovanovic (@vuule)
  - Jason Lowe (@jlowe)

URL: https://github.com/rapidsai/cudf/pull/7084