Fast path single column sort (#7167)
This change is based on the changes in PR #7075. When `cudf::sort()` or `cudf::sorted_order()` is called with a `cudf::table_view` and specifies only a single strings column, we choose a fast-path sort algorithm with a simpler comparator specifically coded for string compares. The specialized code path was added to `cudf::sorted_order()` which is called by the other libcudf sort functions. For example, `cudf::sort()` calls `cudf::sorted_order()` and the calls `cudf::gather()` on the input `cudf::table_view()` to materialize the results. The libcudf `sorted_order` feature has two APIs: `cudf::sorted_order()` and `cudf::stable_sorted_order()` which internally use `thrust::sort()` and `thrust::stable_sort()` respectively. Each uses the `row_lexographic_comparator` for managing sort of multiple columns. A simpler comparator can be used in the case of a single column per the implementation in #7075.

In this PR, I generalized this fast-path for other single column types. I found the same comparator from #7075, templated by type, could be used for speeding up sorting of any comparable type -- where a single column is specified. Further, there are some conditions with numeric types when a comparator is not required and where the `cub::DeviceRadixSort` functions can be used instead of thrust.

The restrictions to account for when _not using a comparator_:
- the type must support an assignment operator as well as the compare operators (basically only numeric types)
- the column must not contain nulls since these are handled specially with a `null_order` parameter
- `thrust::sort()` and `thrust::stable_sort()` sort the input data in-place and do not support descending order
- `cudf::DeviceRadixSort` does not sort in-place and does not have stable-sort but does have a descending order option

Here is how these are used in `cudf::detail::sorted_order<stable>()` matching conditions with these restrictions.

| stable | nulls | numeric | ascending | function |
|:---:|:---:|:---:|:---:| --- |
| y | y | - | - |  `thrust::stable_sort()` with comparator |
| y | - | n | - | `thrust::stable_sort()` with comparator |
| y | - | - | n | `thrust::stable_sort()` with comparator |
| y | n | y | y | `thrust::stable_sort_by_key()` with input column copied |
| n | y | - | - | `thrust::sort()` with comparator |
| n | - | n | - | `thrust::sort()` with comparator |
| n | n | y | y | `cub::DeviceRadixSort::SortPairs` with input column copied and output indices copied |
| n | n | y | n | `cub::DeviceRadixSort::SortPairsDescending` with input column copied and output indices copied |

The `sort_benchmarks.cu` was updated to include a non-nulls set of tests to show the speedups for the bottom half of the chart. The benchmark sorts integers in ascending order. With nulls, the sort is now 1.2x faster. With no nulls, the sort is about 14x faster. The faster speed comes at the expense of 2-3 times the memory required for `thrust::stable_sort_by_key()` or the `cub:DeviceRadixSort::SortPairs()` functions.

The generalization using the new single-column comparator accounts for strings columns as well. So the strings-specific code for this has been removed in this PR.

Authors:
  - David (@davidwendt)

Approvers:
  - Jake Hemstad (@jrhemstad)
  - Karthikeyan (@karthikeyann)

URL: https://github.com/rapidsai/cudf/pull/7167