Fixing empty null lists throwing explode_outer for a loop. (#7649)
I found two issues, one was that we didn't build the correct number of null or empty offsets. We should build them for the exploded column and they are sized as such, but I was marching over it the size of the child data column. This didn't cause trouble as long as there was more data than nulls.

The second issue was the large loop. We have to go over that loop at least the number of nulls we have as that loop is doing two things at once.

1) writing the valid data rows to a gather map.
2) filling in the holes in the gather map for null and empty entries.

This was another case of things working fine as long as we ran the loop enough to cover all the null entries, which happens unless there are more nulls than entries. That wasn't tested and so it was never seen. Thankfully, @sperlingxx tested exactly that.

Added a test for this case.

Fixes #7636

Authors:
  - Mike Wilson (@hyperbolic2346)

Approvers:
  - Nghia Truong (@ttnghia)
  - MithunR (@mythrocks)

URL: https://github.com/rapidsai/cudf/pull/7649