Support setting to a new column in `DataFrame.loc` (#8012)
closes #7628 

This PR adds support to setting a column in the dataframe when the provided column name is a new column name. The specified rows can be of a single row label, a collection of row labels, or slices. The value-to-set can be column-like object or scalar. E.g. you can now do this:

```
>>> x = cudf.DataFrame()
>>> x.loc[:, "a"] = [1, 2, 3] # set a new column with list
>>> x
   a
0  1
1  2
2  3
>>> x.loc[[1, 2], "b"] = ["abc", "cba"] # set part of the new column with list
>>> x
   a     b
0  1  <NA>
1  2   abc
2  3   cba
>>> x.loc[:, "c"] = 5 # set the new column to the scalar
>>> x
   a     b  c
0  1  <NA>  5
1  2   abc  5
2  3   cba  5
```

Authors:
  - Michael Wang (https://github.com/isVoid)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/8012