Remove 10.2 workarounds in groupby functions for dictionary column types (#7949)
Reference #7913 
This removes the compiler restrictions in the libcudf groupby code that was disabled because of compiler issues on nvcc 10.2 when adding dictionary columns support. The corresponding disabled gtests have been enabled as well.

Details of the compiler bug are documented here: https://nvbugswb.nvidia.com/NvBugs5/SWBug.aspx?bugid=3186317

This also fixes an issue where the sum-of-squares aggregation was not working for values specified as a dictionary column.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Karthikeyan (https://github.com/karthikeyann)
  - Nghia Truong (https://github.com/ttnghia)

URL: https://github.com/rapidsai/cudf/pull/7949