Creates an empty column for the null `LIST Scalar` (#8173)
This PR is to support creating a null `LIST Scalar` with an empty column but containing all the necessary children, from a `HostColumnVector.DataType`.

Also removed the function `cudf::jni::make_empty_column`, which is a duplicate of `cudf::empty_like`. and replaced it with the `cudf::empty_like` when building a column from a list scalar. 

closes https://github.com/rapidsai/cudf/issues/8170

Signed-off-by: Firestarman <firestarmanllc@gmail.com>

Authors:
  - Liangcai Li (https://github.com/firestarman)

Approvers:
  - Robert (Bobby) Evans (https://github.com/revans2)
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/8173