Make java rolling window operation APIs consistent (#8251)
There were some recent changes to the rolling window java APIs that made it so there were multiple different ways to set preceding and following and they were inconsistent with each other. What is more if you used the wrong one, you got either the wrong answer or your program crashed with a seg fault trying to dereference a null pointer.

This removes the ambiguity, but it is a breaking change. Because of this I am going to keep this in draft until Monday.

Authors:
  - Robert (Bobby) Evans (https://github.com/revans2)

Approvers:
  - Jason Lowe (https://github.com/jlowe)

URL: https://github.com/rapidsai/cudf/pull/8251